Implement basic function returns
User function calls now are just syntax sugar for setting the arguments, injecting the body and setting a return variable.

This leads to worse error messages which will need to be worked on in the future, but most importantly `return` does not break execution.

Need to implement goto / `Skip` which is gonna be quite annoying since overwatch only allows relative jumps.

But now functions should be practically "fully featured"