Rewrite tokenizer, parser
* Tests will now be located in the /src/ directory
* Tokenizer and parser hand written rather than relying on untyped, potentially slow peggy.
* Remove old tests
* Rewrite code for biome lints
* Remove js in imports for modified files

Basically rewriting the entire compiler besides the type solver which I think is good still.

Pretty much, I don't want to rely on parser generators anymore since it's actually pretty simple to make your own parser with the help of regex for the tokenizer and recursive descent for the parser. Also, I found a very nice workflow for dealing with tagged union type data.

This is nowhere near usable, the analyzer will likely need to be mostly rewritten as well.